# src/main/resources/contract/api.yaml
openapi: 3.0.0
info:
  title: Proyecto DAW - Empresa de Tatuajes
  description: API para manejar el back-end de una empresa de tatuajes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /Subcategory:
    get:
      summary: Obtener todas las subcategorías
      description: "Permite a cualquier usuario obtener la lista de subcategorías disponibles."
      operationId: getSubcategories
      tags:
        - Subcategories
      responses:
        "200":
          description: Lista de subcategorías obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                title: SubcategoriesResponse
                properties:
                  subcategories:
                    type: array
                    description: Lista de subcategorías disponibles
                    items:
                      $ref: "#/components/schemas/Subcategory"
        "404":
          description: No se encontraron subcategorías
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear una nueva subcategoría
      description: "Permite a los administradores crear una nueva subcategoría proporcionando los datos requeridos."
      operationId: createSubcategory
      tags:
        - Subcategories
      security:
        - bearerAuth: []
      requestBody:
        description: Datos de la nueva subcategoría
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subcategory"
      responses:
        "201":
          description: Subcategoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
        "400":
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Subcategory/{idSubcategory}:
    get:
      summary: Obtener una subcategoría por ID
      description: "Permite a cualquier usuario obtener los datos de una subcategoría específica mediante su ID único."
      operationId: getSubcategoryById
      tags:
        - Subcategories
      parameters:
        - name: idSubcategory
          in: path
          required: true
          description: ID único de la subcategoría a obtener
          schema:
            type: integer
      responses:
        "200":
          description: Subcategoría obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
        "404":
          description: Subcategoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar una subcategoría por ID
      description: "Permite a los administradores actualizar los datos de una subcategoría específica."
      operationId: updateSubcategory
      tags:
        - Subcategories
      security:
        - bearerAuth: []
      parameters:
        - name: idSubcategory
          in: path
          required: true
          description: ID único de la subcategoría a actualizar
          schema:
            type: integer
      requestBody:
        description: Nuevos datos para la subcategoría
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subcategory"
      responses:
        "200":
          description: Subcategoría actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
        "400":
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Subcategoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar una subcategoría por ID
      description: "Permite a los administradores eliminar una subcategoría específica mediante su ID único."
      operationId: deleteSubcategory
      tags:
        - Subcategories
      security:
        - bearerAuth: []
      parameters:
        - name: idSubcategory
          in: path
          required: true
          description: ID único de la subcategoría a eliminar
          schema:
            type: integer
      responses:
        "204":
          description: Subcategoría eliminada exitosamente
        "404":
          description: Subcategoría no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subcategory:
      type: object
      properties:
        id_subcategory:
          type: integer
          description: "Identificador único de la subcategoría"
        id_category:
          type: integer
          description: "Identificador de la categoría a la que pertenece esta subcategoría"
        name:
          type: string
          description: "Nombre de la subcategoría"
        description:
          type: string
          description: "Descripción de la subcategoría"




    # Productos actualizados (no tocar)
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Código del error
        errorMessage:
          type: string
          description: Descripción del error
security:
  - bearerAuth: []